fix some spacing and typos

+++ Update-OSX.sh
@@ -1,3 +1,11 @@
+#!/bin/bash
+#Check for sufficient privileges
+if [ "$EUID" -ne 0 ]; then
+  echo "Please run with Elevated Privileges"
+  read -rsp $'Press any key to continue . . .\n' -n1 key
+  exit
+fi
+echo
 echo Updating and Re-installing VisualStudio Code Extensions for webOS
 echo
 git reset --hard
